# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minha API",
        "description": "Descrição da API",
        "contact": {
            "email": "contato@exemplo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/users": {
            "get": {
                "summary": "Lista todos os usuários",
                "description": "Retorna uma lista de todos os usuários cadastrados",
                "operationId": "ec4fe381075eaaa20ceb989f2b968eff",
                "responses": {
                    "200": {
                        "description": "Lista de usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Armazena um novo usuário",
                "description": "Recebe os dados do formulário e cria um novo usuário no banco de dados",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso"
                    },
                    "400": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/users/{name}": {
            "get": {
                "summary": "Exibe um usuário específico",
                "description": "Retorna os detalhes de um usuário com base no nome fornecido",
                "operationId": "c4a9b9643b2ae7168ba52e147adc06b5",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Nome do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        },
        "/users/create": {
            "get": {
                "summary": "Exibe o formulário de criação de usuário",
                "description": "Retorna a view com o formulário para adicionar um novo usuário",
                "operationId": "bb980692892d97fb9e0a7c42914a8670",
                "responses": {
                    "200": {
                        "description": "Formulário de adição de usuário"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "name",
                    "cpf",
                    "data_nascimento"
                ],
                "properties": {
                    "id": {
                        "description": "ID do usuário",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "cpf": {
                        "description": "CPF do usuário",
                        "type": "string"
                    },
                    "data_nascimento": {
                        "description": "Data de nascimento do usuário",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        }
    }
}